+++ SmartPower/package/CHANGELOG.md
@@ -1,6 +1,7 @@
-# v1.2 (pre-release)
+# v1.2 (19 Jul 2023)
 * [Update] Change `deDE` strings.
-* [Fix #2] Allow specifying if the engine must be generating power regardless to the smart logic.
+* [Enhancement #1] Add a setting to control batteries charging at per engine level.
+* [Enhancement #2] Allow specifying if the engine must be generating power regardless to the smart logic.
 
 # v1.1 (14 Jul 2023)
 * [Enhancement] Add `deDe` locale file. Thanks to juf0816@Discord.